+++ src/modele/CoursAvancer.java
@@ -2,7 +2,7 @@ package modele;
 
 public class CoursAvancer extends Cours{
 
-	public CoursAvancer(int annee, int nbHeure, String intituler) {
+	public CoursAvancer(int annee, Double nbHeure, String intituler) {
 		super(annee, nbHeure, intituler);
 	}
