+++ src/modele/CoursIntermediare.java
@@ -1,7 +1,7 @@
 package modele;
 
 public class CoursIntermediare extends Cours {
-	public CoursIntermediare(int annee, int nbHeure, String intituler) {
+	public CoursIntermediare(int annee, Double nbHeure, String intituler) {
 		super(annee, nbHeure, intituler);
 	}
 }
